+++ src/App.js
@@ -7,7 +7,7 @@ import Rooms from './pages/Rooms';
 import AddReservation from './components/AddReservation';
 import Login from './pages/Login';
 import SignUp from './pages/SignUp';
-import MyReservations from './pages/MyReservations';
+import MyReservations from './components/MyReservations';
 
 function App() {
   const isLogged = () => {
