+++ src/components/MyReservations.css
@@ -1,0 +1,9 @@
+.my-reservations {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+#reservation {
+  padding: 1rem;
+}
