+++ src/index.css
@@ -28,6 +28,6 @@ code {
 
 .reservation {
   background-color: aqua;
-  width:100%;
+  width: 100%;
   height: 100vh;
 }
